+++ compiler/wasm/build.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+# tinygo build -o contract.wasm -target wasi ./
+#GOMAXPROCS=1 TINYGO_DISABLE_SIGNAL_STACK=1 tinygo build -o contract.wasm -target wasi ./
+GOMAXPROCS=1 TINYGO_DISABLE_SIGNAL_STACK=1 tinygo build -o contract.wasm -target=wasi -opt=z -no-debug ./
