+++ go.mod
@@ -3,13 +3,16 @@ module github.com/govm-net/vm
 go 1.23.0
 
 require (
+	github.com/spf13/cobra v1.8.0
 	github.com/stretchr/testify v1.8.4
-	github.com/tetratelabs/wazero v1.9.0
+	github.com/tetratelabs/wazero v1.6.0
 	golang.org/x/text v0.23.0
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
